Fix: Injector's re-indexing of existing documents (#16)
* Fix: Injector's re-indexing of existing documents

Resolved an issue with the Injector application. Previously, if a document existed in a prior index, the Injector would incorrectly index it in the current day's index, leading to duplicated data. Now, the Injector correctly updates the document in the prior index without duplicating it in the current day's index.

* add logger

* fix loop break logical

---------

Co-authored-by: htrmind7 <han.tran@mind7.com>